+++ Sources/SystemNotification/View+SystemNotification.swift
@@ -10,48 +10,54 @@ import SwiftUI
 
 public extension View {
     
-    /**
-     Attach a ``SystemNotification`` to the view.
-     
-     View-based system notifications give you control over a
-     single notification.
-     
-     After attaching the modifier, you can use the `isActive`
-     binding to present the provided `content`.
-     */
+    /// Attach a ``SystemNotification`` to the view.
+    ///
+    /// State-based system notifications make it easy to use
+    /// a single binding to control a specific notification.
+    ///
+    /// After applying the modifier, you can use the binding
+    /// to present the provided `content`.
+    func systemNotification<Content: View>(
+        isActive: Binding<Bool>,
+        content: @escaping () -> Content
+    ) -> some View {
+        ZStack {
+            self
+            SystemNotification(
+                isActive: isActive,
+                content: { _ in content() }
+            )
+        }
+    }
+    
+    @available(*, deprecated, message: "Use the new view modifiers to apply style and configuration. These have no effect anymore!")
     func systemNotification<Content: View>(
         isActive: Binding<Bool>,
         configuration: SystemNotificationConfiguration = .standard,
         style: SystemNotificationStyle = .standard,
         content: @escaping () -> Content
     ) -> some View {
-        ZStack(alignment: style.edge.alignment) {
+        ZStack {
             self
             SystemNotification(
                 isActive: isActive,
-                configuration: configuration,
-                style: style,
                 content: { _ in content() }
             )
         }
     }
     
-    /**
-     Attach a system notification context to the view.
-     
-     Context-based system notifications make it easy to show
-     many different notifications with a single context.
-     
-     After attaching the modifier, you can use the context's
-     present functions to present different notifications.
-     */
+    /// Attach a system notification context to the view.
+    ///
+    /// Context-based system notifications make it very easy
+    /// to show multiple notifications with a single context.
+    ///
+    /// After applying the modifier, you can use the context
+    /// to present notifications.
     func systemNotification(
         _ context: SystemNotificationContext
     ) -> some View {
         self.systemNotification(
             isActive: context.isActiveBinding,
-            configuration: context.configuration,
-            style: context.style,
             content: { context.content }
         )
     }